toolchain: remove CLANG_PACKAGES_LTO
thanks to the _IS_CONFIGURE environment variable in the compiler and linker wrappers, which skips LTO during configure, ThinLTO no longer causes any failures or slows down the build. Instead, it speeds up compilation and provides a fast linker cache. So there is no longer any reason to disable it, and it should always be enabled